requirements/dev.txt
-r base.txt
# ── Dev ───────────────────────────────────────────────────────────────────────
django-debug-toolbar
django-extensions